Expose FakeTimeThreadPool
